requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Django==3.2.13
django-admin-list-filter-dropdown==1.0.3
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.8.4
django-filebrowser==3.14.3
django-filter==21.1
django-grappelli==2.15.5
django-redis==5.2.0
django-rq==2.5.1
djangorestframework==3.13.1
drf-extensions==0.7.1
packaging==21.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
rq==1.10.1
six==1.16.0
sqlparse==0.4.2
wrapt==1.14.1
